backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytz>=2023.3
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1